python routines are added, with proper __init__.py files.
