--- bindings/pylibROM/linalg/__init__.py
@@ -1,7 +1,0 @@
-from .pyVector import Vector
-from .pyMatrix import Matrix 
-from .pyBasisWriter import BasisWriter
-from .pyBasisReader import BasisReader
-from .pyBasisGenerator import BasisGenerator 
-from .pyOptions import  Options 
-from . import svd
