--- bindings/pylibROM/linalg/svd/__init__.py
@@ -1,3 +1,0 @@
-from .pySVD import SVD
-from .pyStaticSVD import StaticSVD 
-from .pyIncrementalSVD import IncrementalSVD
