+++ bindings/pylibROM/mfem/__init__.py
@@ -1,1 +1,2 @@
-# from .Utilities import ComputeCtAB
+# __init__.py is needed only for pure python routines.
+from .Utilities import ComputeCtAB
