--- bindings/pylibROM/utils/__init__.py
@@ -1,2 +1,0 @@
-from .mpi_utils import split_dimension, get_global_offsets
-from .pyDatabase import Database
