+++ tests/test_pyIncrementalSVD.py
@@ -1,9 +1,16 @@
 import pytest
 import sys
-sys.path.append("../build")
-
-import pylibROM.linalg as libROM
-import pylibROM.linalg.svd as SVD 
+try:
+    # import pip-installed package
+    import pylibROM
+    import pylibROM.linalg as libROM
+    import pylibROM.linalg.svd as SVD 
+except ModuleNotFoundError:
+    # If pip-installed package is not found, import cmake-built package
+    sys.path.append("../build")
+    import _pylibROM as pylibROM
+    import _pylibROM.linalg as libROM
+    import _pylibROM.linalg.svd as SVD 
 import numpy as np 
 
 def test_getSpatialBasis():
